requirements.txt
presidio-analyzer
presidio-anonymizer
spacy
customtkinter
pyperclip
pystray
Pillow
pytest
